+++ Changelog.md
@@ -4,6 +4,7 @@
 - --test-output option added to [*sams-collector*](docs/sams-collector.md).
 - Collectd output module added.
 - Enables to use --show-software and --show-path together in [*sams-software-updater*](docs/sams-software-updater.md).
+- Export software with user/project specific information in software/version/local version. [*backend/SoftwareAccounting*](docs/backend/SoftwareAccounting.md)
 
 # Version 1.2 - 2020-04-06